% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{compute_bc_day}
\alias{compute_bc_day}
\title{Compute BC Day}
\usage{
compute_bc_day(calendar_year)
}
\arguments{
\item{calendar_year}{An integer representing the calendar year.}
}
\value{
A \code{Date} object representing BC Day.
}
\description{
This function computes BC Day for a given calendar year.
BC Day is defined as the first Monday in August.
}
\examples{
compute_bc_day(2025)
}
