% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{compute_easter_monday}
\alias{compute_easter_monday}
\title{Compute Easter Monday}
\usage{
compute_easter_monday(calendar_year)
}
\arguments{
\item{calendar_year}{An integer representing the calendar year.}
}
\value{
A \code{Date} object representing Easter Monday.
}
\description{
This function computes the date of Easter Monday for a given calendar year. It is defined as
the day following Easter Sunday.
}
\examples{
compute_easter_monday(2025)
}
