% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{compute_easter_sunday}
\alias{compute_easter_sunday}
\title{Compute Easter Sunday using the Meeus/Jones/Butcher Algorithm}
\usage{
compute_easter_sunday(calendar_year)
}
\arguments{
\item{calendar_year}{An integer representing the calendar year.}
}
\value{
A \code{Date} object representing Easter Sunday in the given year.
}
\description{
This function computes the date of Easter Sunday for a given calendar year using the
Meeus/Jones/Butcher algorithm.
}
\examples{
compute_easter_sunday(2025)
}
