% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{compute_family_day}
\alias{compute_family_day}
\title{Compute Family Day}
\usage{
compute_family_day(calendar_year)
}
\arguments{
\item{calendar_year}{An integer representing the calendar year.}
}
\value{
A \code{Date} object representing Family Day.
}
\description{
This function computes Family Day for a given calendar year.
Family Day is defined as the third Monday in February.
}
\examples{
compute_family_day(2025)
}
