% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{compute_good_friday}
\alias{compute_good_friday}
\title{Compute Good Friday}
\usage{
compute_good_friday(calendar_year)
}
\arguments{
\item{calendar_year}{An integer representing the calendar year.}
}
\value{
A \code{Date} object representing Good Friday.
}
\description{
This function calculates Good Friday for the specified calendar year.
Good Friday is defined as two days before Easter Sunday.
}
\examples{
compute_good_friday(2025)
}
