% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{compute_labour_day}
\alias{compute_labour_day}
\title{Compute Labour Day}
\usage{
compute_labour_day(calendar_year)
}
\arguments{
\item{calendar_year}{An integer representing the calendar year.}
}
\value{
A \code{Date} object representing Labour Day.
}
\description{
This function computes Labour Day for a given calendar year.
Labour Day is defined as the first Monday in September.
}
\examples{
compute_labour_day(2025)
}
