% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{compute_thanksgiving}
\alias{compute_thanksgiving}
\title{Compute Thanksgiving Day}
\usage{
compute_thanksgiving(calendar_year)
}
\arguments{
\item{calendar_year}{An integer representing the calendar year.}
}
\value{
A \code{Date} object representing Thanksgiving Day.
}
\description{
This function computes Thanksgiving Day for a given calendar year.
In this context, Thanksgiving is defined as the second Monday in October.
}
\examples{
compute_thanksgiving(2025)
}
