% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-geocode-data.R
\name{get_geocode_with_region}
\alias{get_geocode_with_region}
\title{Geocode Addresses and Attach Regional Information}
\usage{
get_geocode_with_region(
  address,
  api_key = NULL,
  faults = TRUE,
  rate_limit = 1000,
  coords = c("longitude", "latitude"),
  crs = 4326
)
}
\arguments{
\item{address}{A character vector of addresses to geocode.}

\item{api_key}{Optionally, an API key for accessing the BC Address Geocoder API.
Defaults to `NULL`.}

\item{faults}{Logical. If `TRUE`, the function returns flagged issues for each address
(if any). The result will include a second `data.frame` named `faults`. Defaults to `TRUE`.}

\item{rate_limit}{Maximum number of requests allowed per minute. Defaults to 1000.
Use `NULL` for no rate limit.}

\item{coords}{A character vector specifying the names of the columns holding geographic coordinates.
Defaults to `c("longitude", "latitude")`.}

\item{crs}{An integer specifying the Coordinate Reference System (CRS) of the input data.
Defaults to `4326` (WGS 84).}
}
\value{
If `faults = TRUE`, returns a list with two `data.frame`s:
\describe{
  \item{\code{data}}{The geocoded data with attached regional information.}
  \item{\code{faults}}{Any issues noted by the geocoder tool (e.g., addresses that could not be geocoded).}
}
If `faults = FALSE`, only the geocoded data with attached regional information is returned as a `data.frame`.
}
\description{
A convenience function that geocodes addresses using the BC Address Geocoder API
and attaches regional boundary data based on the resulting latitude and longitude.
}
\details{
This function combines \code{\link{get_geocode_data}} and
\code{\link{spatial_join_with_boundaries}} into a single workflow. It first geocodes
addresses to obtain latitude and longitude coordinates using the BC Address Geocoder API.
Then, it performs a spatial join to attach regional data
to the geocoded results.
}
\examples{
\dontrun{
addresses <- c("601 W Broadway, Vancouver, BC", "1081 Burrard St, Vancouver, BC")

## Geocode addresses and attach regional data
addresses_geocoded <- get_geocode_with_region(addresses)

## Access the geocoded data
geo_data <- addresses_geocoded$data

## Access any faults (if faults = TRUE)
faults_data <- addresses_geocoded$faults
}

}
