% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{get_holidays}
\alias{get_holidays}
\title{Get Official Government Statutory Holidays}
\usage{
get_holidays(calendar_year)
}
\arguments{
\item{calendar_year}{An integer representing the calendar year.}
}
\value{
A vector of \code{Date} objects representing the official holidays.
}
\description{
This function returns a vector of official government statutory holidays
(without weekend adjustments) for the specified calendar year. It includes
both fixed-date holidays and variable-date holidays computed by corresponding functions.
}
\examples{
get_holidays(2025)

}
