% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{get_observed_holidays}
\alias{get_observed_holidays}
\title{Get Observed Government Statutory Holidays}
\usage{
get_observed_holidays(calendar_year)
}
\arguments{
\item{calendar_year}{An integer representing the calendar year.}
}
\value{
A vector of \code{Date} objects representing the observed holidays.
}
\description{
This function returns a vector of observed government statutory holiday dates for a given calendar year.
Official holiday dates that fall on a weekend are adjusted to the following Monday.
}
\examples{
get_observed_holidays(2025)

}
