% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{is_holiday}
\alias{is_holiday}
\title{Check if a Date is an Official Holiday}
\usage{
is_holiday(query_date)
}
\arguments{
\item{query_date}{A \code{Date} object to be evaluated.}
}
\value{
\code{TRUE} if the date is an official holiday; \code{FALSE} otherwise.
}
\description{
This function checks whether the given query date matches any official government statutory holiday.
}
\examples{
is_holiday(as.Date("2025-07-01"))

}
