% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{is_observed_holiday}
\alias{is_observed_holiday}
\title{Check if a Date is an Observed Holiday}
\usage{
is_observed_holiday(query_date)
}
\arguments{
\item{query_date}{A \code{Date} object to be evaluated.}
}
\value{
\code{TRUE} if the date is an observed holiday; \code{FALSE} otherwise.
}
\description{
This function checks whether the given query date is an observed government statutory holiday,
taking weekend adjustments into account.
}
\examples{
is_observed_holiday(as.Date("2025-07-01"))

}
