% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-holidays.R
\name{observed_date}
\alias{observed_date}
\title{Compute Observed Holiday Date}
\usage{
observed_date(holiday_date)
}
\arguments{
\item{holiday_date}{A \code{Date} object representing the original holiday date.}
}
\value{
A \code{Date} object representing the observed holiday date.
}
\description{
This helper function calculates the observed date for a given holiday.
If the holiday falls on a weekend (Saturday or Sunday), the observed date is shifted to the following Monday.
}
\examples{
observed_date(as.Date("2025-12-25"))
}
